fix(utils): separate test utils from regular utils to prevent npm from trying to load dev dependenci
